src/valkey_stress_test/scenarios/loader.py
"""YAML scenario file parser."""

from __future__ import annotations

import yaml
from pathlib import Path
from typing import Dict, Any
import logging
from .models import Scenario

logger = logging.getLogger(__name__)


class ScenarioLoader:
    """Loads and parses scenario definitions from YAML files."""
    
    def __init__(self):
        """Initialize scenario loader."""
        # TODO: Initialize validation schema
        pass
        
    def load_scenario(self, scenario_path: Path) -> Scenario:
        """Load a scenario from YAML file."""
        # TODO: Read YAML file
        # TODO: Validate structure
        # TODO: Create Scenario object
        raise NotImplementedError()
        
    def validate_scenario(self, scenario_data: Dict[str, Any]) -> bool:
        """Validate scenario structure and values."""
        # TODO: Check required fields
        # TODO: Validate step definitions
        # TODO: Verify workload references
        raise NotImplementedError()
        
    def load_builtin_scenario(self, name: str) -> Scenario:
        """Load a built-in scenario by name."""
        # TODO: Map name to builtin scenario file
        # TODO: Load and return scenario
        raise NotImplementedError()
