pyproject.toml
[tool.poetry]
name = "valkey-stress-test"
version = "0.1.0"
description = "Memory stress testing tool for Valkey-Search with vector operations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "valkey_stress_test", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.0"
redis = "^5.0.0"
psutil = "^5.9.0"
prometheus-client = "^0.19.0"
typer = "^0.9.0"
pyyaml = "^6.0"
h5py = "^3.10.0"
aiofiles = "^23.0"
pandas = "^2.0.0"
rich = "^13.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
vst = "valkey_stress_test.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
